TEC-4768 move URNs into packages
