+++ src/main/java/com/echobox/api/linkedin/types/locale/LocaleURN.java
@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 
-package com.echobox.api.linkedin.types;
+package com.echobox.api.linkedin.types.locale;
+
+import com.echobox.api.linkedin.types.URN;
 
 import lombok.Getter;
